[File] Move stackedBaseName() into file folder and add unit test
